refactor: standardize store events
