+++ services/storageService.ts
@@ -1,9 +1,7 @@
-import { Platform } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { Platform } from 'react-native';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { get, set, del, keys } from 'idb-keyval';
-import logger from '@/utils/logger';
+import logger from '../utils/logger';
 
 
 export class StorageService {
